requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
google-generativeai>=0.8.0
protobuf>=5.26.0

